Refactoring - Removing duplicate PluralForms from German and Portuguese
